Move error_map.go to new errors package. Fix tests.
